Fix inability to customize IInfo/Info classes.
There was a race condition that would cause an ImportError if you tried
to extend schema.I<name>Info or schema.<name>Info in a model class'
module. The solution in this commit is to create the schema classes for
model, IInfo and Info before importing or creating the concrete classes.