Base component status on /Status w/o meta_type.
Zenoss platform behavior is currently (as of 5.0.2) to derive every
component's status property from the events under the
/Status/<component-meta_type> event class. Using a child event class
under /Status has long been deprecated as a best practice as the event's
component field already conveys the component context.

This change allows events in and under the /Status event class and
associated to the component to affect the component's status property.

Refs ZEN-17536.