Fix __init__.py references
The two places I modified reference things defined in __init__.py, but all the __init__.py files seem to be largely empty.  I am guessing this is stuff that was defined in __init__.py in the original (non-ZPL) tutorial, but is now defined in zenpack.yaml.  I modified it to read how I think it should, but if it's wrong, feel free to reject the PR.