Fixes modeler timeout issue
The problem is limited to components with indexed attributes.  In
certain cases the modeler times out due to the following:

1) index scope too wide for catlog objects since ZPL will always choose
'Products.ZenModel.DeviceComponent.DeviceComponent' objects in the
catalog (all components) instead of just objects of the relevant
component class
2) presence of bad catalog brain objects will cause modeler tracebacks
if getObject() fails, which also causes modeler to fail.

- Fixed by narrowing focus of catalog results scope to relevant class
entries
- Added commented code to uncatalog bad entries if appropriate to do so
here