Sequence graphs/graphpoints by definition order
Previously all graphs and graphpoints were getting the default sequence
value of 0. This resulted in the user not having control over the order
that graphs and their graphpoints would be shown.

Now the sequence of each graph and graphpoint will be set according to
the order they appear within the template and graph (respectively) in
zenpack.yaml

Fixes ZEN-23590.