Device status is now critical event(s) in /Status
This seems like the simplest way to think of status that applies to as
many different types of devices as possible. If you want to mark your
device as "down", create a critical severity event for the device in, or
under, the /Status event class.

Devices that are not being monitored because their production state is
less than zProdStateThreshold will still have an unknown status.