Issue warning for using zenoss keywords
Fixes ZEN-19460

Using a python keyword is an error, but some words, like description or
sysUpTime, should just be a warning.  When creating datasource/datapoints,
sysUpTime can be pretty common so no need to warn on them.  If the word is
a param for the spec class then don't worry about it.  Updated unit test to
account for this update