Because 'seen' was tracking objects, rather than the way that the object was reached, there was potential to filter out objects that could be reached through multiple paths out, before we got to the proper path to it (the one specified in extra_paths).
I changed the 'seen' set to track the way that the object is
reached (source, relationship, dest) instead, which accomplishes
the same thing (avoiding cycles in the data), but does so more
correctly.

I also added a maximum depth, as there are often a crazy number
of bogus and insane paths through the model when recurse_all
is being used.