Revert adding WinService YAML definitions
Fixes ZEN-25733

Removing Windows Service definitions for now.  We'll keep feature/ZEN-25733
branch in case we want to add it in the future.  We should make it more
desirable for users to use the WinService templates and datasources as
opposed to the WinService page.  Competing WinService definitions across
multiple ZenPacks could also cause confusion and/or overwrite any desired
setups for service monitoring.