Reword the getting started section for ZenPacks.zenoss.ZenPackLib
Existing text still pointed the user to downloading a single zenpacklib.py
to include in their ZenPack.  This makes it clear that zpl is now in a
ZenPack which needs to be installed.  Changed Downloading to denote Legacy.

TODO: Once the new website is ready we'll need to point the user to the
ZenPack download area.  Also, will it be packaged with core?  If so, we should
say so.